requirements.txt
streamlit
sentence-transformers
spacy
pdfminer.six
python-docx
pytesseract
scikit-learn
pyresparser
pandas
plotly
